mailing/requirements.txt
astroid==2.8.2
backcall==0.2.0
click==8.0.3
colorama==0.4.4
debugpy==1.5.0
decorator==5.1.0
entrypoints==0.3
Flask==2.0.2
gunicorn==20.1.0
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
isort==5.9.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
jupyter-client==7.0.6
jupyter-core==4.8.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
nest-asyncio==1.5.1
parso==0.8.2
pickleshare==0.7.5
platformdirs==2.4.0
prompt-toolkit==3.0.20
Pygments==2.10.0
pylint==2.11.1
python-dateutil==2.8.2
python-dotenv==0.19.1
python-http-client==3.3.3
# pywin32==302
pyzmq==22.3.0
sendgrid==6.8.2
six==1.16.0
starkbank-ecdsa==2.0.0
toml==0.10.2
tornado==6.1
traitlets==5.1.0
typing-extensions==3.10.0.2
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.12.1
